XC-215: split tests
